pyproject.toml
[project]
name = 'shiftings'
version = '0.2.0'
description = 'Simple shift management system'
readme = 'README.md'
requires-python = '>=3.9'
authors = [
    { name = 'HaDiKo e.V. HaDiNet', email = 'software@hadiko.de' },
]
classifiers = [
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3 :: Only',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
]

dependencies = [
    'Django==4.0.9',
    'django-bootstrap5==22.2',
    'django-ical==1.8.3',
    'django-phonenumber-field[phonenumberslite]==6.3.0',
    'django-stubs==1.12.0',
    'icalendar==4.1.0',
    'types-python-dateutil==2.8.19',
    'gunicorn==20.1.0',
    'holidays==0.14.2',
    'Pillow==9.4.0',
    'psycopg2-binary==2.9.3',
    'pylint==2.14.5',
    'python-dateutil==2.8.2',
    'Authlib==1.0.1',
    'django-auth-ldap==4.1.0',
    'requests==2.28.1',
    'types-requests==2.28.3',
    'pytz==2022.1',
    'django-colorfield==0.8.0',
    'mkdocs==1.4.2',
    'mkdocs-material==9.0.11'
]

[project.optional-dependencies]
tests = [
    'types-requests==2.28.3',
    'types-click==7.1.2',
    'types-Flask==1.1.6',
    'types-PyYAML==6.0.4',
    'mypy==0.991',
    'pylint==2.14.5',
]

[project.urls]
repository = 'https://github.com/HaDiNet/shiftings'

[tool.setuptools.packages.find]
where=['src']

[tool.setuptools.package-data]
shiftings = ['py.typed']
